ofmcc-1602 - move Reporting View to Reports folder
