+++ backend/src/components/lookup.js
@@ -3,7 +3,7 @@ const { getOperation, getLabelFromValue } = require('./utils')
 const HttpStatus = require('http-status-codes')
 const _ = require('lodash')
 const cache = require('memory-cache')
-const { RequestCategoryMappings, FiscalYearMappings, MonthMappings, RequestSubCategoryMappings } = require('../util/mapping/Mappings')
+const { RequestCategoryMappings, RequestSubCategoryMappings, FiscalYearMappings, MonthMappings } = require('../util/mapping/Mappings')
 const { MappableObjectForFront } = require('../util/mapping/MappableObject')
 const log = require('../components/logger')
 
@@ -19,6 +19,17 @@ async function getRequestCategories() {
     lookupCache.put('requestCategories', requestCategories, ONE_HOUR_MS)
   }
   return requestCategories
+}
+
+async function getRequestSubCategories() {
+  let requestSubCategories = lookupCache.get('requestSubCategories')
+  if (!requestSubCategories) {
+    requestSubCategories = []
+    const response = await getOperation('ofm_subcategories?$orderby=ofm_display_order')
+    response?.value?.forEach((item) => requestSubCategories.push(new MappableObjectForFront(item, RequestSubCategoryMappings)))
+    lookupCache.put('requestSubCategories', requestSubCategories, ONE_HOUR_MS)
+  }
+  return requestSubCategories
 }
 
 async function getFiscalYears() {
@@ -41,17 +52,6 @@ async function getMonths() {
     lookupCache.put('months', months, ONE_HOUR_MS)
   }
   return months
-}
-
-async function getRequestSubCategories() {
-  let requestSubCategories = lookupCache.get('requestSubCategories')
-  if (!requestSubCategories) {
-    requestSubCategories = []
-    const response = await getOperation('ofm_subcategories?$orderby=ofm_display_order')
-    response?.value?.forEach((item) => requestSubCategories.push(new MappableObjectForFront(item, RequestSubCategoryMappings)))
-    lookupCache.put('requestSubCategories', requestSubCategories, ONE_HOUR_MS)
-  }
-  return requestSubCategories
 }
 
 async function fetchAndCacheData(cacheKey, operationName) {
